+++ packages/qiankun/src/index.ts
@@ -1,2 +1,4 @@
 export * from './apis/loadMicroApp';
 export * from './apis/registerMicroApps';
+export * from './apis/isRuntimeCompatible';
+export * from './types';