+++ packages/sandbox/src/core/membrane/utils.ts
@@ -1,30 +1,36 @@
-import { defineProperty, getOwnPropertyDescriptor, hasOwnProperty } from '@qiankunjs/shared';
+import { defineProperty, getOwnPropertyDescriptor, getOwnPropertyNames, hasOwnProperty } from '@qiankunjs/shared';
 import { isBoundedFunction, isCallable, isConstructable } from '../../utils';
 
 const functionBoundedValueMap = new WeakMap<CallableFunction, CallableFunction>();
 
-export function getTargetValue<T>(target: unknown, value: T): T {
+export function rebindTarget2Fn<T>(target: unknown, fn: T, receiver: unknown): T {
   /*
     仅绑定 isCallable && !isBoundedFunction && !isConstructable 的函数对象，如 window.console、window.atob 这类，不然微应用中调用时会抛出 Illegal invocation 异常
     目前没有完美的检测方式，这里通过 prototype 中是否还有可枚举的拓展方法的方式来判断
     @warning 这里不要随意替换成别的判断方式，因为可能触发一些 edge case（比如在 lodash.isFunction 在 iframe 上下文中可能由于调用了 top window 对象触发的安全异常）
    */
-  if (isCallable(value) && !isBoundedFunction(value) && !isConstructable(value as CallableFunction)) {
-    const typedValue = value as CallableFunction;
+  if (isCallable(fn) && !isBoundedFunction(fn) && !isConstructable(fn as CallableFunction)) {
+    const typedValue = fn as CallableFunction;
     const cachedBoundFunction = functionBoundedValueMap.get(typedValue);
     if (cachedBoundFunction) {
       return cachedBoundFunction as T;
     }
 
-    const boundValue = Function.prototype.bind.call(typedValue, target) as CallableFunction;
+    const boundValue = function proxyFunction(...args: unknown[]): unknown {
+      return Function.prototype.apply.call(
+        typedValue,
+        target,
+        args.map((arg) => (arg === receiver ? target : arg)),
+      );
+    };
 
-    // some callable function has custom fields, we need to copy the enumerable props to boundValue. such as moment function.
-    // use for..in rather than Object.keys.forEach for performance reason
-    for (const key in typedValue) {
-      ((boundValue as object) as Record<string, unknown>)[key] = ((typedValue as object) as Record<string, unknown>)[
-        key
-      ];
-    }
+    // some callable function has custom fields, we need to copy the own props to boundValue. such as moment function.
+    getOwnPropertyNames(typedValue).forEach((key) => {
+      // boundValue might be a proxy, we need to check the key whether exist in it
+      if (!hasOwnProperty(boundValue, key)) {
+        defineProperty(boundValue, key, getOwnPropertyDescriptor(typedValue, key)!);
+      }
+    });
 
     // copy prototype if bound function not have but target one have
     // as prototype is non-enumerable mostly, we need to copy it from target function manually
@@ -49,20 +55,25 @@ export function getTargetValue<T>(target: unknown, value: T): T {
 
       if (valueHasInstanceToString || boundValueHasPrototypeToString) {
         const originToStringDescriptor = getOwnPropertyDescriptor(
-          valueHasInstanceToString ? value : Function.prototype,
+          valueHasInstanceToString ? fn : Function.prototype,
           'toString',
         );
 
-        Object.defineProperty(boundValue, 'toString', {
-          ...originToStringDescriptor,
-          ...(originToStringDescriptor?.get ? null : { value: () => typedValue.toString() }),
-        });
+        Object.defineProperty(
+          boundValue,
+          'toString',
+          Object.assign(
+            {},
+            originToStringDescriptor,
+            originToStringDescriptor?.get ? null : { value: () => typedValue.toString() },
+          ),
+        );
       }
     }
 
-    functionBoundedValueMap.set(value, boundValue);
+    functionBoundedValueMap.set(fn, boundValue);
     return boundValue as T;
   }
 
-  return value;
+  return fn;
 }