+++ packages/shared/src/error/index.ts
@@ -1,1 +1,0 @@
-export { QiankunError } from './QiankunError';