+++ packages/ui-bindings/vue/.fatherrc.js
@@ -1,0 +1,6 @@
+import cfg from '../../../.fatherrc.cjs';
+
+export default {
+  umd: {},
+  ...cfg,
+};