+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,103 @@
 # @qiankunjs/sandbox
 
+## 0.0.1-rc.12
+
+### Patch Changes
+
+- bd12dbad: fix: defer scripts should wait until html loaded
+- Updated dependencies [bd12dbad]
+  - @qiankunjs/shared@0.0.1-rc.9
+
+## 0.0.1-rc.11
+
+### Patch Changes
+
+- 98b071bf: feat: support defer scripts and keep the executing order to consist with browser
+- Updated dependencies [98b071bf]
+  - @qiankunjs/shared@0.0.1-rc.8
+
+## 0.0.1-rc.10
+
+### Patch Changes
+
+- f2af2e36: feat: extract NodeTransformer type to shared package
+- Updated dependencies [f2af2e36]
+  - @qiankunjs/shared@0.0.1-rc.7
+
+## 0.0.1-rc.9
+
+### Patch Changes
+
+- d3e9872d: feat(sandbox): use cloneNode api instead of importNode for compatible
+- 7cc06bd4: feat(loader): add lru cache for assets fetch by default
+- Updated dependencies [54b0878e]
+- Updated dependencies [7ba95cf2]
+- Updated dependencies [312abbc7]
+- Updated dependencies [6f074136]
+  - @qiankunjs/shared@0.0.1-rc.6
+
+## 0.0.1-rc.8
+
+### Patch Changes
+
+- 43bf37a5: fix(sandbox): should get container from getter function in every accessing
+- a34a92a9: feat(sandbox): micro app mounting should wait unit rebuilding link element loaded to avoid unstyleed content flash
+- 7cf93b54: fix(sandbox): createElement hijack must be paired to avoid rewriting leak
+- 32106b11: fix(sandbox): dynamic async script order should calculate on the fly
+
+## 0.0.1-rc.7
+
+### Patch Changes
+
+- 5f77347b: feat(sandbox): support dynamic sync scripts executed by order in sandbox
+- Updated dependencies [5f77347b]
+- Updated dependencies [8e54e129]
+  - @qiankunjs/shared@0.0.1-rc.5
+
+## 0.0.1-rc.6
+
+### Patch Changes
+
+- 2aca545c: fix: should invoke getContainer method to get container every time to avoid reference misordering
+
+## 0.0.1-rc.5
+
+### Patch Changes
+
+- 3d1d3367: fix: should patch the container head/body element immediately rather than patch its functions with proxy
+
+## 0.0.1-rc.4
+
+### Patch Changes
+
+- 488447ad: ✨ set proxy appendChild/insertBefore method for every sandbox rather than modify prototype on HTMLElement
+- dc4d9aef: 🐛parallel sandbox should use different compartment id
+- e7d788ef: feat: not rebind non-native global properties
+- 76b6bff7: 🐛 compatible with webpack chunk cache logic
+- Updated dependencies [76b6bff7]
+  - @qiankunjs/shared@0.0.1-rc.4
+
+## 0.0.1-rc.3
+
+### Patch Changes
+
+- 39301f19: 🔀 merge master
+  - @qiankunjs/shared@0.0.1-rc.3
+
+## 0.0.1-rc.2
+
+### Patch Changes
+
+- Updated dependencies [b23d3d7b]
+  - @qiankunjs/shared@0.0.1-rc.2
+
+## 0.0.1-rc.1
+
+### Patch Changes
+
+- Updated dependencies [ebb2bcaa]
+  - @qiankunjs/shared@0.0.1-rc.1
+
 ## 0.0.1-beta.6
 
 ### Patch Changes